% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_timemap.R
\name{get_timemap}
\alias{get_timemap}
\title{TimeMap for access to version history}
\usage{
get_timemap(uri = NULL)
}
\arguments{
\item{uri}{desired uri of the web resource}

\item{limit}{limit}
}
\description{
TimeMap for access to version history
}
\examples{
\dontrun{
# mementos
tt <- get_timemap("http://www.base-search.net/about/de/about_sources_date_dn.php?menu=2")
# timemap
tt <- get_timemap("http://cnn.com")
}
}

